Move usePointerEventListener hook into its own file and export TrackDisplayMode
