+++ src/core/Slider/Slider.tsx
@@ -9,49 +9,9 @@ import '@itwin/itwinui-css/css/slider.css';
 import { TooltipProps } from '../Tooltip';
 import { CommonProps } from '../utils/props';
 import { getBoundedValue } from '../utils/common';
+import { usePointerEventListener } from '../utils/hooks/usePointerEventListener';
 import { Track } from './Track';
 import { Thumb } from './Thumb';
-
-// based on https://usehooks.com/useEventListener/
-function usePointerEventListener(
-  eventName: string,
-  handler: (event: PointerEvent) => void,
-  element: HTMLElement | Document | undefined,
-) {
-  // Create a ref that stores handler
-  const savedHandler = React.useRef<(event: PointerEvent) => void>();
-
-  // Update ref.current value if handler changes.
-  // This allows our effect below to always get latest handler ...
-  // ... without us needing to pass it in effect deps array ...
-  // ... and potentially cause effect to re-run every render.
-  React.useEffect(() => {
-    savedHandler.current = handler;
-  }, [handler]);
-
-  React.useEffect(
-    () => {
-      // Make sure element supports addEventListener
-      const isSupported = element && element.addEventListener;
-      if (!isSupported) {
-        return;
-      }
-
-      // Create event listener that calls handler function stored in ref
-      const eventListener = (event: PointerEvent) =>
-        savedHandler.current?.(event);
-
-      // Add event listener
-      element?.addEventListener(eventName, eventListener);
-
-      // Remove event listener on cleanup
-      return () => {
-        element?.removeEventListener(eventName, eventListener);
-      };
-    },
-    [eventName, element], // Re-run if eventName or element changes
-  );
-}
 
 /**
  * Determines which segments are shown with color.
